Drive : added a function to remove any rotation of the wheels AngleCorrection : all the angle are now in degrees, convertion done when calling the function + simplification of the function

+++ MicroController/EE579WS/ToyCar/lib/modules/angleCorrection.h
@@ -1,6 +1,6 @@
 #ifndef ANGLECOR_H
 #define ANGLECOR_H
 
-void angleCorrection(float currentAngle, float targetAngle);
+void angleCorrection(int currentAngle, int targetAngle);
 
 #endif
